feat: 채팅 목록에서 채팅방으로의 내비게이션 연동
- ChattingRoute.kt: 채팅방 이동을 위한 `ChatRoomNavKey` 및 `ChatListNavKey` 정의
- ChatSummary.kt: `chatRoomId` 필드 타입을 `Int`에서 `Long`으로 변경
- ChatRoomScreen.kt: 채팅방 화면 구성을 위한 `ChatRoomScreen` 컴포저블 신규 생성
- ChattingEntryBuilder.kt: `chattingEntry`에 `ChatRoomNavKey` 진입점 추가 및 화면 이동을 위한 콜백 파라미터 정의
- ChatListScreen.kt: 채팅 아이템 클릭 이벤트를 처리할 수 있도록 `onChatRoomClick` 콜백 구현 및 데이터 바인딩
- AppScreen.kt: `AppNavigator`를 통해 실제 채팅방 이동(`navigate`) 및 뒤로가기(`popBackStack`) 로직 연결
- AppNavigator.kt: 불필요한 주석 제거 및 내비게이션 관련 TODO 문구 정리

+++ core/domain/src/main/java/kr/co/call/domain/model/chatting/ChatSummary.kt
@@ -12,7 +12,7 @@ package kr.co.call.domain.model.chatting
  *
  */
 data class ChatSummary(
-    val chatRoomId: Int = -1, // 매니저 챗방은 -1
+    val chatRoomId: Long = -1, // 매니저 챗방은 -1
     val image: String = "", // 매니저의 경우에는 없음
     val name: String = "",
     val isMainCharacter: Boolean = false,
